requirements.txt
python-dotenv
paho-mqtt
influxdb-client[ciso]
systemd
